feat: 일정 조율 결과 순위 뷰에서 list item에 Link 연결
