+++ frontend/src/features/discussion/ui/DiscussionDetailLink/index.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const linkStyle = style({
+  width: '100%',
+});
