+++ frontend/src/features/discussion/ui/DiscussionRank/index.css.ts
